Duplicate code in readMetadata.js & versionFallback.js #725
