+++ lib/rename-version.js
@@ -12,6 +12,7 @@ const fs = require('fs');
 const glob = require('glob');
 const path = require('path');
 const readMetadata = require('./server/readMetadata.js');
+const utilsMetadata = require('./server/utilsMetadata.js');
 
 const CWD = process.cwd();
 
@@ -99,7 +100,7 @@ if (fs.existsSync(CWD + '/versioned_docs/version-' + currentVersion)) {
     if (extension !== '.md' && extension !== '.markdown') {
       return;
     }
-    const res = readMetadata.extractMetadata(fs.readFileSync(file, 'utf8'));
+    const res = utilsMetadata.extractMetadata(fs.readFileSync(file, 'utf8'));
     const metadata = res.metadata;
     const rawContent = res.rawContent;
     if (!metadata.id) {
