+++ lib/version.js
@@ -14,6 +14,7 @@ const mkdirp = require('mkdirp');
 const chalk = require('chalk');
 const readMetadata = require('./server/readMetadata.js');
 const versionFallback = require('./server/versionFallback.js');
+const utilsMetadata = require('./server/utilsMetadata.js');
 const env = require('./server/env.js');
 
 const CWD = process.cwd();
@@ -78,7 +79,7 @@ files.forEach(file => {
     return;
   }
 
-  const res = readMetadata.extractMetadata(fs.readFileSync(file, 'utf8'));
+  const res = utilsMetadata.extractMetadata(fs.readFileSync(file, 'utf8'));
   let metadata = res.metadata;
   // Don't version any docs without any metadata whatsoever.
   if (Object.keys(metadata).length === 0) {
